test: migrate useAsyncCacheFetch to vitest
